# docs/swagger.yaml
definitions:
...
  models.TaskEvent:
    properties:
      actor_id:
        type: integer
      created_at:
        type: string
      event_id:
        type: string
      event_type:
        type: string
      id:
        type: integer
      payload:
        type: object
      project_id:
        type: integer
      task_id:
        type: integer
      task_version:
        type: integer
    type: object
...
  realtime.ProjectLock:
    properties:
      field:
        type: string
      holder_user_id:
        type: integer
      holder_username:
        type: string
      task_id:
        type: integer
    type: object
  realtime.ProjectPresence:
    properties:
      connections:
        type: integer
      user_id:
        type: integer
      username:
        type: string
    type: object
  realtime.ProjectServerMessage:
    properties:
      cursor:
        type: integer
      error:
        type: string
      event:
        $ref: '#/definitions/models.TaskEvent'
      event_id:
        type: string
      events:
        items:
          $ref: '#/definitions/models.TaskEvent'
        type: array
      has_more:
        type: boolean
      lock:
        $ref: '#/definitions/realtime.ProjectLock'
      next_cursor:
        type: integer
      presence:
        items:
          $ref: '#/definitions/realtime.ProjectPresence'
        type: array
      project_id:
        type: integer
      server_node_id:
        type: string
      type:
        type: string
    type: object
...
  response.Resp:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /projects/{id}/ws:
    get:
      consumes:
      - application/json
      description: 'Upgrades to a project WebSocket stream for PROJECT_INIT, PROJECT_SYNC,
        TASK_CREATED, TASK_UPDATED, TASK_DELETED, PRESENCE_SNAPSHOT, TASK_LOCKED,
        TASK_UNLOCKED, LOCK_ERROR, PING, and PONG messages. Authenticate with Authorization:
        Bearer <token> or the token query parameter.'
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer JWT
        in: header
        name: Authorization
        type: string
      - description: JWT when a WebSocket client cannot send Authorization
        in: query
        name: token
        type: string
      - description: Last seen task event id
        in: query
        name: cursor
        type: integer
      - description: Alias of cursor
        in: query
        name: last_event_id
        type: integer
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols; subsequent frames use this message schema
          schema:
            $ref: '#/definitions/realtime.ProjectServerMessage'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/response.Resp'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/response.Resp'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/response.Resp'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/response.Resp'
        "503":
          description: WebSocket hub is not configured
          schema:
            $ref: '#/definitions/response.Resp'
      summary: Project realtime WebSocket
      tags:
      - Realtime
...
swagger: "2.0"